Primera fase de migración a Postgres. Los cambios fueron principalmente a nivel de los units tests en sí mismos.

+++ querying-based-on-similarity/src/fuzzy/operations/DropFuzzyDomainOperation.java
@@ -24,9 +24,14 @@ public class DropFuzzyDomainOperation extends Operation {
 
     @Override
     public void execute() throws SQLException {
+        if (this.connector.getCatalog().equals("")) {
+            throw new SQLException("No database selected");
+        }
+        String schemaName = this.connector.getCatalog();
+
         Logger.debug("Starting DROP FUZZY DOMAIN " + domain + " operation");
         String sql = "DELETE FROM information_schema_fuzzy.domains " +
-            "WHERE table_schema = (select database()) " + 
+            "WHERE table_schema = '"+schemaName+"'" + 
             "AND domain_name = '" + domain + "'";
         
         int rows = connector.fastUpdate(sql);
